+++ js/main.js
@@ -1,4 +1,4 @@
-import {getPosts} from './util.js';
+import {getPosts} from './data.js';
 
 // eslint-disable-next-line no-console
 console.table(getPosts(25));
